Refactor: useUtterance 분리
