+++ src/components/Comment/index.tsx
@@ -1,31 +1,9 @@
-import { useEffect, useRef } from 'react'
-
-const customUtteranceAttribute = {
-  src: 'https://utteranc.es/client.js',
-  repo: 'kimyouknow/kimyouknow.github.io',
-  'issue-term': 'pathname',
-  theme: 'github-light',
-  label: '💬 comments',
-  crossorigin: 'anonymous',
-  async: 'true',
-}
+import useUtterance from '@/hooks/useUtterance'
 
 const Comment = () => {
-  const element = useRef<HTMLDivElement>(null)
+  const { utteranceElement } = useUtterance()
 
-  useEffect(() => {
-    if (element.current === null) return
-    const utterances: HTMLScriptElement = document.createElement('script')
-    const attributes = customUtteranceAttribute
-
-    Object.entries(attributes).forEach(([key, value]) => {
-      utterances.setAttribute(key, value)
-    })
-
-    element.current.appendChild(utterances)
-  }, [])
-
-  return <div ref={element} />
+  return <div ref={utteranceElement} />
 }
 
 export default Comment
